+++ client/src/TakeSurvey.tsx
@@ -77,8 +77,10 @@ const TakeSurvey: React.FC = () => {
 	const [submitted, setSubmitted] = useState(false);
 	const [loading, setLoading] = useState(false);
 	const [error, setError] = useState('');
-	const [assessmentResults, setAssessmentResults] = useState<AssessmentResult[]>([]);
+    const [assessmentResults, setAssessmentResults] = useState<AssessmentResult[]>([]);
 	const [scoringResult, setScoringResult] = useState<ScoringResult | null>(null);
+    // For one-question-per-page: gate questions behind an initial personal info step
+    const [infoStepDone, setInfoStepDone] = useState(false);
 
 	// Enable anti-cheating measures for assessments and quizzes
 	const isAssessmentType = survey && TYPES_REQUIRING_ANSWERS.includes(survey.type);
@@ -126,7 +128,7 @@ const TakeSurvey: React.FC = () => {
 		setQuestionsLoaded(true);
 	};
 
-	useEffect(() => {
+    useEffect(() => {
 		// If slug is provided, fetch that specific survey
 		if (slug) {
 			setLoading(true);
@@ -149,7 +151,7 @@ const TakeSurvey: React.FC = () => {
 						);
 					}
 
-					setSurvey(res.data);
+                    setSurvey(res.data);
 
 					// For manual surveys, load questions immediately
 					// For question bank surveys, wait for user email
@@ -166,7 +168,16 @@ const TakeSurvey: React.FC = () => {
 			// Otherwise fetch all surveys for selection
 			axios.get<Survey[]>('/api/surveys').then(res => setSurveys(res.data));
 		}
-	}, [slug]);
+    }, [slug]);
+
+    // Reset info step when survey or navigation mode changes
+    useEffect(() => {
+        if (survey?.navigationMode === NAVIGATION_MODE.ONE_QUESTION_PER_PAGE) {
+            setInfoStepDone(false);
+        } else {
+            setInfoStepDone(true);
+        }
+    }, [survey?.navigationMode, survey?._id]);
 
 	const handleAnswerChange = (qid: string, value: string) => {
 		setForm({ ...form, answers: { ...form.answers, [qid]: value } });
@@ -185,6 +196,22 @@ const TakeSurvey: React.FC = () => {
 			loadQuestions(survey, email);
 		}
 	};
+
+    const canStart = () => {
+        const hasName = form.name && form.name.trim().length > 0;
+        const hasEmail = form.email && form.email.includes('@');
+        return Boolean(hasName && hasEmail);
+    };
+
+    const handleStart = async () => {
+        if (!survey) return;
+        if (!canStart()) return;
+        // Ensure questions are loaded for question bank surveys before starting
+        if (survey.sourceType === SOURCE_TYPE.QUESTION_BANK && form.email && !questionsLoaded) {
+            await loadQuestions(survey, form.email);
+        }
+        setInfoStepDone(true);
+    };
 
 	const handleSubmit = async (e: React.FormEvent) => {
 		e.preventDefault();
@@ -209,9 +236,62 @@ const TakeSurvey: React.FC = () => {
 
 				const results: AssessmentResult[] = questions.map(q => {
 					const userAnswer = form.answers[q._id];
-					const correctAnswer =
-						q.correctAnswer !== undefined ? q.options[q.correctAnswer] : '';
-					const isCorrect = userAnswer === correctAnswer;
+
+					// Proper answer matching logic
+					let isCorrect = false;
+					let correctAnswerText = '';
+
+					if (q.correctAnswer !== undefined && userAnswer !== undefined) {
+						if (q.type === 'single_choice') {
+							// For single choice, correctAnswer is an index
+							if (typeof q.correctAnswer === 'number') {
+								const userOptionIndex = q.options?.findIndex(opt =>
+									typeof opt === 'string' ? opt === userAnswer : opt.text === userAnswer
+								);
+								isCorrect = userOptionIndex === q.correctAnswer;
+
+								// Get correct answer text
+								const correctOption = q.options?.[q.correctAnswer];
+								correctAnswerText = typeof correctOption === 'string'
+									? correctOption
+									: correctOption?.text || '';
+							} else {
+								// Fallback for direct text comparison
+								isCorrect = userAnswer === q.correctAnswer;
+								correctAnswerText = String(q.correctAnswer);
+							}
+						} else if (q.type === 'multiple_choice' && Array.isArray(q.correctAnswer)) {
+							// Handle multiple choice
+							const userAnswerArray = Array.isArray(userAnswer) ? userAnswer : [userAnswer];
+							const userOptionIndices = userAnswerArray
+								.map(ans =>
+									q.options?.findIndex(opt =>
+										typeof opt === 'string' ? opt === ans : opt.text === ans
+									)
+								)
+								.filter(idx => idx !== -1);
+							const correctIndices = q.correctAnswer as number[];
+							isCorrect =
+								userOptionIndices.length === correctIndices.length &&
+								userOptionIndices.every(idx => correctIndices.includes(idx));
+
+							// Get correct answer text
+							correctAnswerText = correctIndices
+								.map(idx => {
+									const option = q.options?.[idx];
+									return typeof option === 'string' ? option : option?.text || '';
+								})
+								.join(', ');
+						} else if (q.type === 'short_text') {
+							// For short text, compare directly
+							isCorrect = userAnswer === q.correctAnswer;
+							correctAnswerText = String(q.correctAnswer);
+						} else {
+							// Fallback logic
+							isCorrect = userAnswer === q.correctAnswer;
+							correctAnswerText = String(q.correctAnswer);
+						}
+					}
 					const maxPoints =
 						q.points ||
 						survey.scoringSettings?.customScoringRules?.defaultQuestionPoints ||
@@ -232,7 +312,7 @@ const TakeSurvey: React.FC = () => {
 						questionText: q.text,
 						descriptionImage: q.descriptionImage,
 						userAnswer: userAnswer || '',
-						correctAnswer: correctAnswer,
+						correctAnswer: correctAnswerText,
 						isCorrect,
 						pointsAwarded,
 						maxPoints,
@@ -535,7 +615,7 @@ const TakeSurvey: React.FC = () => {
 								</div>
 							</div>
 
-							{questionsLoaded ? (
+                            {questionsLoaded && (survey.navigationMode !== NAVIGATION_MODE.ONE_QUESTION_PER_PAGE || infoStepDone) ? (
 								// Conditional rendering based on navigation mode
 								survey.navigationMode === NAVIGATION_MODE.ONE_QUESTION_PER_PAGE ? (
 									<OneQuestionPerPageView
@@ -752,7 +832,7 @@ const TakeSurvey: React.FC = () => {
 										))}
 									</div>
 								)
-							) : survey?.sourceType === SOURCE_TYPE.QUESTION_BANK ? (
+                            ) : survey?.sourceType === SOURCE_TYPE.QUESTION_BANK && (survey.navigationMode !== NAVIGATION_MODE.ONE_QUESTION_PER_PAGE || infoStepDone) ? (
 								<div className='text-center py-8'>
 									<div className='text-purple-500 text-6xl mb-4'>🎲</div>
 									<h3 className='text-xl font-semibold text-gray-700 mb-2'>
@@ -765,6 +845,20 @@ const TakeSurvey: React.FC = () => {
 									</p>
 								</div>
 							) : null}
+
+                            {/* Start button for one-question-per-page before entering questions */}
+                            {survey?.navigationMode === NAVIGATION_MODE.ONE_QUESTION_PER_PAGE && !infoStepDone && (
+                                <div className='flex justify-center pt-4'>
+                                    <button
+                                        type='button'
+                                        onClick={handleStart}
+                                        disabled={!canStart()}
+                                        className='btn-primary px-8 py-3 disabled:opacity-50 disabled:cursor-not-allowed'
+                                    >
+                                        Start
+                                    </button>
+                                </div>
+                            )}
 
 							{/* Submit button - only show for non-one-question-per-page modes */}
 							{survey?.navigationMode !== NAVIGATION_MODE.ONE_QUESTION_PER_PAGE && (
