+++ routes/admin/shared/middleware.js
@@ -1,0 +1,9 @@
+const asyncHandler = require('../../../middlewares/asyncHandler');
+const { jwtAuth } = require('../../../middlewares/jwtAuth');
+const imageUpload = require('../../../middlewares/imageUpload');
+
+module.exports = {
+  asyncHandler,
+  jwtAuth,
+  imageUpload,
+};
